Shuffle to simplify AccessPath deps.
